gyb: alter windows path for filename substitutions
When building on Windows, the filepath contains `\` as the path
separator.  Substitute `/` for the path separator prior to replacement.
This allows clang to properly compile the source file.  Windows supports
both as a path separator and this matches what clang does by default for
the filepaths when preprocessing.  This allows gyb generated files to
build on Windows.