Replace default implementation for _hash(into:) with automatic synthesis
This removes the glaring cheat of force-synthesizing _hash(into:) whenever we derive hashValue.

_hash(into:) is now automatically derived whenever it's not explicitly implemented, including in classes and extensions of (possibly imported or synthesized) structs, enums or classes.

This is not entirely free of magic, though. To determine which implementation to generate, the body synthesizer takes a peek into the parent decl context -- if it finds an implicit hashValue there, the generated body implements hashing from scratch, feeding components into the hasher. Otherwise, we assume hashValue has been resolved to an explicit implementation, and we generate the body for the (previously) default implementation of _hash(into:).

This change introduces some diagnostic regressions, and also requires a followup change in the CoreFoundation overlay.