Implement heterogeneous floating-point comparisons, remove old compare methods.
This patch makes two fairly significant changes to the FloatingPoint and BinaryFloatingPoint protocols and concrete types.

- Provides heterogeneous comparison between members conforming to BinaryFloatingPoint. This was part of the original SE-0067, but was not implementable at the time because we didn't have SE-0104 yet.

- Makes the `isEqual(to:)`, `isLess(than:)` and `isLessThanOrEqualTo()` methods on FloatingPoint and concrete types obsolete. At the time that these were added, this was planned to be the means of conforming to Equatable / Comparable, but we ended up going in a different direction with the static operators ==, <, <= instead.

Of particular relevance, the second change is breaking, I don't think that much code uses these methods directly, but if people have types that conform to FloatingPoint or BinaryFloatingPoint, they may have used these as the implementation hooks, as that was the original documented design. I'm not sure if we can actually make this change, and if we can I'm not precisely sure what affordances we need to make, so feedback on this point is particularly welcome.