Teach deabstraction how to handle top level code in Playgrounds, Scripts, REPL.
The hard thing about this is that variables and constants get turned into
global variables representationally, and that functions defined in that program
can access and mutate those variables.  To handle this, we have the compiler
"forcibly flatten" the code down into a single function, which guarantees we
can see all the accesses.  We then promote the globals to stack allocations and
then to SSA.

This isn't the ideal approach for various reasons (code size can explode, there
are some cases that probably are not handled right) but is good enough to
handle the common cases, and should carry us through March.

This also fixed deabstraction to only print its dump when it is modifies a function, make it a  lot less annoying to consume on the command line.

PiperOrigin-RevId: 187894451