[stdlib] Revisit underestimatedCount default implementation for Collection
Currently the default implementation of `underestimatedCount` property
for a `Collection` is to simply return `count`. Problem with that is
`count` is allowed to be O(n) operation on some collections and can have
a side-effect of iterating over the whole collection, which defeats the
whole purpose of `underestimatedCount`.

The proposed change makes `underestimatedCount` strictly O(1), which
requires a certain pessimization of it's return value for `Collection`
instances that are only forward or bidirectional.
`RandomAccessCollection` instances guarantee constant time `count` and
therefore still can provide a default implementation of
`underestimatedCount` in terms of `count`.