[+0-all-args] For staging in purposes, add in a lit feature to distinguish in between tests that must run with a +0/+1 runtime.
This will enable me to have a small commit to enable +0 and the ability to
easily revert if needed. It makes sense to do this since +0 is a disruptive
change and the number of tests modified is not /that/ big (i.e. ~200).

This shouldn't last for more than a week or two.

rdar://34222540