Fix obvious typo. SWIFT_BULID_STDLIB => SWIFT_BUILD_STDLIB.
I think we didn't catch this since we also check if overlays are enabled here
and generally overlays and stdlib are enabled together. It is NFC from the
actual build's perspective since this just controls whether or not diagnostic
information is printed out about what SDKs/etc is configured to be built.