Change check-swift-linux-x86_64 to use ninja instead of the sed pipeline
This is a hacky fix for an issue where build-script-impl gets the ninja
command invocations from `ninja -n -v` incorrectly. Ninja gets a few
commands, including `clang++ ... -o somelib`, which assumes incorrectly
that we're in `swift-linux-x86_64` and thus errors out on the build
process.

This change just lets ninja do the work at the cost of potentially
scrambling the lit progress bar.