[stdlib] optimise DefaultIndices.contains because it is sorted.
Currently DefaultIndices uses Sequence.contains, which starts from the beginning and checks equality with each index until it finds a match and returns true, or reaches the end and returns false. But the indices of collections are by definition sorted. This commit adds a `contains(_ element: Element)` method which returns false immediately if ‘element’ is not less than endIndex. It then goes through each index from the beginning and returns false as soon as it reaches an index that is larger than ‘element’.