[test] Make sure directory exists before putting a module cache in it
This test uses %swiftc_driver_plain instead of %swiftc_driver or
%target-build-swift because it's testing what the driver does when
invoked with no -swift-version argument, but that means we have to
manually set up things like a module cache path. This will then fail
if it's the first Driver test run in a clean build directory (because
of where lit.py decides to put temporary files).

rdar://problem/38354843