[stdlib] Set, Dictionary: Enable per-instance hash seeding
This is done by xoring the base address of the hash table storage to the hash seed.

In deterministic mode, we perturb the seed by the bucket count instead, so that we ordering within hashed collections remains repeatable (as long as the capacity doesn't change).

Fixes https://bugs.swift.org/browse/SR-3268