Add DifferentiationPrimalValues, `primal(for:)` and `adjoint(for:with:backpropagating:)` as requirements to DifferentiableModule.
Making the primal produce a struct containing primal values and the adjoint take such a struct is important for avoiding recomputation (or, forced checkpointing :-)). The struct acts like a static tape. This is mimicking the SIL-level implementation of AD, and should all go away when AdjointGen handles generating a partial-producing function for each intermediate function application.

With this addition, `applied(to:)` and `gradient(for:backpropagating:)` become default implementations. When AD is in, `applied(to:)` will become a requirement again, and all of `primal(...)`, `gradient(...)` and `adjoint(...)` will be nuked.

Separately, a convenience initializer `init(filterShape:strides:padding:)` is added.

PiperOrigin-RevId: 188980846