SIL: Stored property initializers in non-resilient modules might need public linkage
Because of a ill-advised core team decision, Swift now
allows @_inlineable root initializers on types that
are not @_fixed_layout (aa85e4512fa42e23562f6f41e4ad3343143631a0).

Since stored property initializers of types that are not
@_fixed_layout are permitted to reference non-public symbols,
they cannot be serialized and must instead have public
linkage. This negates most of the benefit of inlining the
constructor in the first place, but oh well.

This does not affect resilient builds.

Fixes <rdar://problem/38439363>.