Test tgmath without printing (#15218)
* Test tgmath without printing

The tgmath tests have been comparing the printed form of the numeric results.
This worked only because the old floating-point formatting logic was inaccurate.
As such, they implicitly tolerate slight inaccuracies.

This commit rewrites the test to verify the numeric values directly.
This is both more explicit and more robust.  (In particular, it
protects this test against future changes to the floating-point
formatting logic.)

For example, the previous code verified `log1p` with this code:
```
(d1, f1, g1) = (log1p(dx), log1p(fx), log1p(gx))
expectEqual("log1p 0.0953101798043249 0.0953102 log1p",
  print3("log1p", d1, f1, g1))
```
which implicitly relies on inaccurate printing.

The new code uses a custom helper:
```
expectEqualWithAccuracy(log1p(dx), log1p(fx), log1p(gx), 0.09531017980432487, ulps: 1)
```
which verifies that the `Double`, `Float`, and `CGFloat` values
match the provided number within 1 ULP.

Note: I've somewhat arbitrarily set the `ulps:` parameter to 1 for most tests.
This might be unnecessarily strict and these parameters could be relaxed: The
goal of the tgmath test is simply to ensure that Swift's functions map to the
correct C math library routines.  This test is not trying to validate the
accuracy of those underlying routines.

* Verify that C `nextafter` matches Swift `.nextUp`/`.nextDown`

As suggested by Steve Canon.

* Verify only the bottom three bits of remquo().1

Note: The C/C++ standards only require the bottom three
bits to be correct, although macOS assures 7 bits.

* Comment and simplify a bit

Note: it's possible that this commit will cause test failures on platforms other than Mac and Linux/x86 if libm is not sufficiently accurate on these values. Previously these tests were enforcing that values had to print identically, now there is an explicit tolerance. The new approach is formally better, but it may be necessary to relax some of the tolerances from 1ulp to 2 or more for some platforms.