Fix a bug in deabstraction/propagateTensorOperand which caused it to crash on bb arguments created by switch_enum terminators.  This came up with checking optionals.
This allows re-enabling the no-copy test, but the test now requires copies
because of the optional unwrap that we can't deabstract yet, so move it to the
integration test.

Also, adopt llvm::PrettyStackTrace in deabstraction so we get more symbolic
context if/when a test fails in g3.

I'm still confused why mean() doesn't return a scalar in these cases though, it
seems like it should?  If it did, then this would work fine.

PiperOrigin-RevId: 189142837