Implement basic SGD optimizer.
This again exposed many constraints in our static system. The `update(with:by:)` method should be overloaded for many aggregates instead of just one. We want to be able to write:

  parameters.update(with: grads, moments, by: { param, grad, moment in
    ...
  })

While the short answer is "just add overloads!", this is not elegant or scalable, and will make the ParameterAggregate protocol explode. We need to revisit the aggregate parameter updates once again after the demo to find other ways.

PiperOrigin-RevId: 189198494