Prevent associated type witnesses from referring to Objective-C type parameters.
Objective-C type parameters are not recoverable from an instance of
a generic Objective-C class, so we cannot refer to them anywhere in
type metadata. However, they could slip in via associated type
inference, which bypassed the existing checks. This would always crash
later in IRGen; ban it in the type checker so that doesn't happen.

Fixes rdar://problem/34979938.