Teach deabstraction to analyze the uses of tfop's not just the operands of them, to look for memory addresses to promote.  This is a prerequisite to fixing a problem Richard discovered involving BatchNormalizationLayer. Until the second half is fixed, this is mostly NFC.
This does make the partitioner print stuff out unconditionally to stderr when
copys are inserted.  Theses are enough of a nuisance for this phase of the
project that we need better debugging tools that are always on.  This may not
be enough, but it will at least help me diagnose things.

PiperOrigin-RevId: 189419467