Be smarter about optional chaining fixits. Previously if we didn't already know a contextual result type to check for optionality we'd always optimistically offer an optional chaining fixit, which results in many new coders hitting the fix button to add an "?" to an expression that needs a non-optional final value, only to hit the fix button again to add an "!" the next compile now that context is clearer.
Now we make a disjunction to try both ways and get a correct result the first
time. The code turns out to be a bit baroque for long dotted member chains
because we need to stepwise make optional each member along the path because we
don't have the advantage of rearranging the AST to group the whole chain here
inside the constraint system, but we manage it while only making a single
disjunction.