Make tgmath tests generic over TGMath protocol, extend to Float80 (#15289)
* Make tgmath tests generic over a TGMath protocol, extend them to Float80.

Define most of the test machinery in terms of a internal TGMath protocol that refines BinaryFloatingPoint. Add a little bit of gyb support to automate adding additional types, as it is likely these tests will need to support Float16 and Float128 in the future. Make all the input values exactly representable in all of the types in question, to avoid representation errors, loosen default tolerance to 3 to make porting to new platforms less noisy.

* Further tgmath test cleanup.

Merge #if conditions where possible so that they're not all over the file. Get rid of placeholder CGFloat on non-Darwin, which is no longer needed and doesn't work with the new approach anyway.

* Remove ldexp, it's been renamed scalbn.