[benchmark] Add a new utility script called Benchmark_QuickCheck.
This benchmark script is similar to the guard malloc/runtime runner, but it only
runs the tests. The intention is that one can use this to quickly in a
multithreaded way verify that all benchmarks run successfully. In contrast, the
normal driver will run only single threaded since it is meant to test
performance, so is not able to take advantage of all cores on a system.

I wrote this quickly to verify some benchmark tests still worked. No point in
not sharing with everyone else.