Fix b/75505813 Partitioner assertion on invalid source location, which happened because we were trying to run the partitioner on the public_external symbols deserialized from the standard library.
Finish implementation of __tf_tensor_from_scalars_1d lowering in deabstraction,
which isn't useful because the constant folder doesn't do anything helpful yet.

Fix the 'implicit copy' diagnostics to be aware of when they are about to
diagnose an invalid location like the above.  In addition to "not crashing",
have them go above and beyond by printing the demangled name of the symbol
they're reporting the warning in, which will make it easier to track down
future similar bugs.  For example, we now get the second line here:

<unknown>:0: warning: value implicitly copied to the accelerator, use .toDevice() to make transfer explicit
<unknown>:0: error: located in TensorProtocol<>._adjointDot(_:partial:seed:) aka '$S10TensorFlow0A8ProtocolPAAs7Numeric6ScalarRpzrlE11_adjointDot_7partial4seedx_xtx_xxtF'

PiperOrigin-RevId: 189656607